mess-auth/requirements.txt
fastapi~=0.109.2
sqlalchemy~=2.0.25
alembic~=1.13.1
python-jose[cryptography]~=3.4.0
passlib[bcrypt]~=1.7.4
pydantic~=2.5.3
uvicorn[standard]~=0.26.0
python-multipart~=0.0.19
python-dotenv~=1.0.1
pydantic-settings~=2.1.0
psycopg2~=2.9.9
asyncpg~=0.29.0
aiosqlite~=0.19.0
